WIP: Get balance from backend
Refactor OfflineBalance to use BigInteger (represent balance in wei)
Has introduced some bugs in withdrawal. Has also disabled the rewarding of currency.

+++ app/src/main/java/com/bakkenbaeck/toshi/presenter/WithdrawPresenter.java
@@ -168,7 +168,7 @@ public class WithdrawPresenter implements Presenter<WithdrawActivity> {
         }
         final Intent intent = new Intent();
         intent.putExtra(INTENT_WALLET_ADDRESS, this.activity.getBinding().walletAddress.getText().toString());
-        intent.putExtra(INTENT_WITHDRAW_AMOUNT, Double.valueOf(this.activity.getBinding().amount.getText().toString()));
+        intent.putExtra(INTENT_WITHDRAW_AMOUNT, Long.valueOf(this.activity.getBinding().amount.getText().toString()));
         this.activity.setResult(RESULT_OK, intent);
         this.activity.finish();
         this.activity.overridePendingTransition(R.anim.enter_from_left, R.anim.exit_to_right);
